chore: rename OperatorCard, OperatorFilters, OperatorGrid, and OperatorTable components to match naming conventions
