+++ apps/web/src/components/operators/OperatorCard.tsx
