+++ apps/web/src/components/operators/OperatorTable.tsx
